Finished all of the request classes for routes for default permissions and conditions checks for their corresponding buttons.

+++ app/Models/Access/Permission/Permission.php
@@ -74,14 +74,18 @@ class Permission extends Model {
 	 * @return string
 	 */
 	public function getEditButtonAttribute() {
-		return '<a href="'.route('admin.access.roles.permissions.edit', $this->id).'" class="btn btn-xs btn-primary"><i class="fa fa-pencil" data-toggle="tooltip" data-placement="top" title="' . trans('crud.edit_button') . '"></i></a>';
+		if (access()->can('edit-permissions'))
+			return '<a href="'.route('admin.access.roles.permissions.edit', $this->id).'" class="btn btn-xs btn-primary"><i class="fa fa-pencil" data-toggle="tooltip" data-placement="top" title="' . trans('crud.edit_button') . '"></i></a>';
+		return '';
 	}
 
 	/**
 	 * @return string
 	 */
 	public function getDeleteButtonAttribute() {
-		return '<a href="'.route('admin.access.roles.permissions.destroy', $this->id).'" class="btn btn-xs btn-danger" data-method="delete"><i class="fa fa-times" data-toggle="tooltip" data-placement="top" title="' . trans('crud.delete_button') . '"></i></a>';
+		if (access()->can('delete-permissions'))
+			return '<a href="'.route('admin.access.roles.permissions.destroy', $this->id).'" class="btn btn-xs btn-danger" data-method="delete"><i class="fa fa-times" data-toggle="tooltip" data-placement="top" title="' . trans('crud.delete_button') . '"></i></a>';
+		return '';
 	}
 
 	/**
